Refactors build scripts to be even more clean. Use compiler "strict-mode" to throw more warnings (not one yet :-D).
Common:
- No longer use "long long", they are not compatible with c++98.

Mutex/SpinLock changes:
- Only use one global header file mutex.h/spinlock.h, instead of one for each implementation.
- Use private-impl idiom to provide a custom impl within each *.cpp.
- Adds linux native mutex and spinlock based on pthread and built-in __sync* functions. This makes it possible to build the library without c++11.

+++ include/humblelogging/appender/rollingfileappender.h
@@ -12,7 +12,7 @@ class HUMBLE_EXPORT_API RollingFileAppender
   : public Appender
 {
 public:
-  RollingFileAppender(const std::string &filename, bool immediate = false, int maxRoll = 5, unsigned long long maxFileSize = 10485760);
+  RollingFileAppender(const std::string &filename, bool immediate = false, int maxRoll = 5, unsigned long maxFileSize = 10485760);
   virtual ~RollingFileAppender();
   virtual void log(const LogEvent &logEvent);
 
@@ -35,7 +35,7 @@ private:
   std::string _filename;
   bool _immediate;
   int _maxRoll;
-  unsigned long long _maxFileSize;
+  unsigned long _maxFileSize;
 
   std::ofstream _stream;
 };
